move tests to folders that reflect the packages of the classes they test
move tests to folders that reflect the packages of the classes they test
