+++ src/test/java/com/salesforce/dataloader/dao/csv/CsvTest.java
@@ -23,7 +23,7 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.salesforce.dataloader.dao;
+package com.salesforce.dataloader.dao.csv;
 
 import java.io.File;
 import java.io.IOException;
